Try to fix cross-version errors
