+++ modules/core/shared/src/main/scala/eu/timepit/refined/macros/LiteralMatchers.scala
@@ -9,29 +9,28 @@ trait LiteralMatchers {
   import c.universe._
 
   private[macros] object BigIntMatcher {
-    private val matcher = Unliftable[Constant] {
-      case q"scala.`package`.BigInt.apply(${lit: Literal})" => lit.value
-    }
-
-    def unapply(expr: c.Tree): Option[BigInt] = matcher.unapply(expr).flatMap { lit =>
-      lit.value match {
-        case i: Int    => Some(BigInt(i))
-        case l: Long   => Some(BigInt(l))
-        case s: String => scala.util.Try(BigInt(s)).toOption
-        case _         => None
-      }
+    def unapply(expr: c.Tree): Option[BigInt] = expr match {
+      case q"scala.`package`.BigInt.apply(${lit: Literal})" =>
+        lit.value.value match {
+          case i: Int    => Some(BigInt(i))
+          case l: Long   => Some(BigInt(l))
+          case s: String => scala.util.Try(BigInt(s)).toOption
+          case _         => None
+        }
+      case _ => None
     }
   }
 
   private[macros] object BigDecimalMatcher {
-    private val matcher = Unliftable[Constant] {
-      case q"scala.`package`.BigDecimal.apply(${lit: Literal})"   => lit.value
-      case q"scala.`package`.BigDecimal.exact(${lit: Literal})"   => lit.value
-      case q"scala.`package`.BigDecimal.valueOf(${lit: Literal})" => lit.value
-    }
+    def unapply(expr: c.Tree): Option[BigDecimal] = {
+      val constant = expr match {
+        case q"scala.`package`.BigDecimal.apply(${lit: Literal})"   => Some(lit.value.value)
+        case q"scala.`package`.BigDecimal.exact(${lit: Literal})"   => Some(lit.value.value)
+        case q"scala.`package`.BigDecimal.valueOf(${lit: Literal})" => Some(lit.value.value)
+        case _                                                      => None
+      }
 
-    def unapply(expr: c.Tree): Option[BigDecimal] = matcher.unapply(expr).flatMap { lit =>
-      lit.value match {
+      constant.flatMap {
         case i: Int    => Some(BigDecimal(i))
         case l: Long   => Some(BigDecimal(l))
         case d: Double => Some(BigDecimal(d))
